a big change to timeout with async stuffs, see below:
so we introduce RC::TIMER which the Timeout middleware would insert into
the env. this is needed for the async clients, where once after the
response is back, it should immediately turn off the timer.

the timer is now using the underlying event framework's timer facility,
because raising exceptions don't really work under event-driven
architecture.

also, the client should be turned off once timeout is reached too.

a high level overview:

under async architecture, either the timer turns off the client,
or the client turns off the timer, depending on which one is done
first.

also, in fiber-aware http client, we can't get the client.
so the client should put itself inside a fiber-local storage,
which is actually Thread.current[...] in ruby...

this is ok because this storage could only be accessed by the
current fiber. one day we might need to change it to Fiber.current[...]
but for now i guess it's fine.