fix a bug with FollowRedirect + Cache + cache.update => true, see:
whenever you use follow redirect + cache + cache.update => true,
it would accidentally cache the previous response body, headers,
and status into the new cache with the new uri, causing it
thinks the old response is the new response, and stuck there.

we need to explicitly say that we want to assign nil when
cache.update is true. cannot rely on RESPONSE_STATUS...