cache.rb: rework how we cache, detail:
so now the default cache.key would also contain the request headers,
we no longer need to special case for oauth1 and oauth2 cache.key,
the regular one with request header in the key should do the trick.

also, we merged status, headers, body cache into one! this way,
we don't have to worry what if one of them are removed by memcache,
but not the other? this would make the internal state of the cache
more consistent to work with.