fix a long standing bug which is hard to track down, would crash EM:
in some cases, if we resume in some of the callbacks,
and raise exceptions (which is very likely to happen in the
application code, isn't it?), the eventmachine would crash!

i am not very sure why, might be a design concept that,
some of the callbacks should never raise an exception,
so the developers should know it as soon as possible.

calling EM.next_tick to resume in another context could
solve this. apparently next_tick is considered a safe callback,
while em-http-request would make call to some unsafe callbacks?