future.rb: well, we need to callback first then resume, or Client#wait
won't wait for the callback. this might not be what we want...
so we callback first then resume. not sure if this would block
something, since the synchronize block is enlarged...
i guess it's ok, since the other synchronize block is merely
trying to wait, and it blocks anyway...

just remember, don't call Client#wait in the callback! that would
definitely cause deadlock for sure

updated: well, no. it won't deadlock, and i don't know why :D
i am willing to find out, but still cannot think of a reason
right now.