client.rb: don't check if the future is loaded and skip waiting, reason:
at first i thought that if the response is ready, there's no point to
wait, and it might cause deadlock when no one is going to awake it.
but whenever we're in ASYNC mode, that is using a callback, then
we *do want* to wait until the callback is done. i did this by wrapping
synchronize{} around setting response and calling callback, thinking
that if Future#wait (Client#wait) is called, then it would first block
by synchronize{}, and then check the response is done, then just leave.

but if we check if the response is ready in Client#wait, then we cannot
be blocked by the synchronize{} block! and it won't hurt if the response
is already ready, because we still have that loaded check in synchronize
block. so this should be the right way to do it.