now eventmachine+thread works. the point is to wrap all callbacks. see:
the point is that we need to use the same strategy (i.e. thread or
fiber) for the requests inside callback, otherwise, we can't easily
make Client#wait work well. since a thread cannot wait for a fiber,
and a fiber cannot wait for a thread either. yes we can let Client#wait
decide how to wait and how to resume, passing down to the future,
but this would be quite complex to implement.

so to make it easier, we can just pick the same strategy, by wrapping
the entire callback. but we cannot only wrap it for the callback,
since Future.create won't be able to find the desired strategy this way.

the problem is that, each eventmachine callback is called in the main
thread! so Future.create would always assume there should be a fiber
wrapped outside, but it would be false if what we really want is a
thread.

this is the best balanced solution for both supporting fibers and threads.