preserve promise and preserve wrapped call
we need to make sure that cached result should be exactly the same as
the original result (without cached). therefore we need to make the
cache middleware work as if an http client, preserving everything from
an http client, i.e. a promise and all the futures, REQUEST_URI, and
also call the callback from the promise as if not cached.

previously, we fake everything... therefore there's no promise at all,
and we call the callback directly after response is back. we can't
return a promise while not calling the callback from the promise,
otherwise response[RESPONSE_BODY] might not be handled from the
callback.

this was discovered by the following code in RC::Github:

    def all path, query={}, opts={}
      q = {:per_page => MAX_PER_PAGE}.merge(query)
      get(path, q, opts.merge(RESPONSE_KEY => PROMISE)).then{ |response|
        # without this commit, the below response[RESPONSE_BODY] would
        # not be handled by the callback, leaving it a string instead of
        # a parsed json when cached.
        body = response[RESPONSE_BODY] + (page_range(response).map{ |page|
          get(path, q.merge(:page => page), opts)
        }.inject(&:+))
        response.merge(RESPONSE_BODY => body)
      }.future_body
    end