so really handle errors in the client instead of ErrorHandler
this could be a radical change that it would raise the exception
from the client instead of from ErrorHandler. now ErrorHandler
only tries to create the error rather than doing both.

this means that clients without ErrorHandler would now raise
exceptions! this should fix the confusion and inconsistent
problem mentioned in #6.

the intention for this change was all for creating RC::Retry,
which could retry the request upon errors. if we raise the exception
from ErrorHandler, then we have no chance to catch the error
before it's raised. this should make RC::Retry much easier to
implement.

hope this won't break too much code!