fix a race condition. detail:
i failed to reproduce this in a smaller case, therefore cannot
write a test for this :( but here's the reproducing code
i am using. just in case i forgot what's the issue.

    RC::Github.pool_size = 5
    g = RC::Github.new
    g.me['html_url']
    u.auth.update_for_github(g)

first, at `me` it would create a worker, printing:

    "signal"
    "waiting = 0"

then the task is done. next, it would enqueue two tasks
at the same time, before it could grab any task:

    "signal"
    "waiting = 1"
    "signal"
    "waiting = 1"

then the worker successfully grabbed a task:

    "grab"

since it would only spawn a new worker when waiting=0,
there would be only one worker doing the job. at the
moment, there's one task in the queue, and one worker
working.

before the task is done, the request is done, and then
a subsequent request would be firing, triggering:

    "signal"
    "waiting = 0"

because there's no waiting workers, a new worker would
be spawned. so now we've two tasks in the queue,
with two workers, one is still working, the other newly
created one would just get a task from the queue.

then we have two workers working, one task in the queue.

next, one of the workers would be done. it got stuck
at this point.

i think i can't fully explain what's going on here...
but another clue is that, suppose then i do this:

    g.me

that is enqueuing another task, then, all tasks could be
done quickly. so the reason should be that all of the
workers are waiting for a task, but none of them are
signalled, while we do still have a task in the queue.

it's really late here. i'll think about this tomorrow.